+++ packages/cspell-lib/src/lib/util/wordSplitter/index.ts
@@ -1,0 +1,2 @@
+export type { SplitOptions, SplitResult } from './wordSplitter.js';
+export { split } from './wordSplitter.js';
